% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linkages.R
\name{linkages}
\alias{linkages}
\title{LINKAGES main function}
\usage{
linkages(iplot, nyear, nspec, fc, dry, bgs, egs, max.ind, plat, temp.mat,
  precip.mat, spp.params, switch.mat, fdat, clat, basesc = 74,
  basesn = 1.64)
}
\arguments{
\item{iplot}{PEcAn settings object}

\item{nyear}{number of years}

\item{nspec}{number of species}

\item{fc}{field capacity}

\item{dry}{wilting point}

\item{bgs}{beginning of growing season}

\item{egs}{end of growing season}

\item{max.ind}{maximum number of indiviuals allowed to grow in a plot}

\item{plat}{latitude}

\item{temp.mat}{monthly temperature matrix (nyear x 12)}

\item{precip.mat}{monthly precipitation matrix (nyear x 12)}

\item{spp.params}{species parameter matrix}

\item{switch.mat}{species switches matrix}

\item{fdat}{underground parameter matrix}

\item{clat}{climate correction factor matrix}

\item{basesc}{initial humus weight}

\item{basesn}{initial N in soil}
}
\value{
several things
}
\description{
Main function for running all LINKAGES subroutines
}
\author{
Ann Raiho
}

